Document the unusual version number
We use an unusual version number. This is done because the version
references the version of Bitcoin Core vendored in `depend` as well as
the version of this library.

Add documentation to the version number so the next dev reading it
doesn't post an issue asking why the unusual format :)