Bump version 0.19.0-3 -> 0.19.0-0.4.0
We recently bumped the MSRV, this is a major change and requires a
version bump.

We are using a version format that includes the Bitcoin Core version and
the version for this lib, currently the lib version is a single number.
Instead use typical semvar <Major>.<Minor>.<Patch> for the lib.