Move bitcoin submodule to tmp/bitcoin
We want to move over to using git subtrees instead of submodules however
we would like to keep the current path `depend/bitcoin` for the subtree.

In preparation for this add a new temporary submodule `tmp/bitcoin` and
checkout the same commit as `depend/bitcoin` so that this commit builds
cleanly.

Remove the subtree at `depend/bitocin`.

Next we will create a subtree at `depend/bitcoin`.