Cut over to using the bitcoin core subtree
We now have a git subtree for the Bitcoin Core code at `depend/bitcoin`.
Cut over from the `tmp/bitcoin` submodule to the `depend/bitcoin`
subtree.

Please note, this moves the bitcoin repository tip to the tagged release of
the 0.19.2 version instead of the tip used in the current submodule (0.19.0
with a few extra patches). We update the library version appropriately
to 0.19.2-0.4.1