Fix Bitcoin Core 0.21.2 upgrade
We just updated the subtree to use branch `0.21.2` using:

git subtree pull --prefix='depend/bitcoin' \
                 git@github.com:bitcoin/bitcoin.git v0.21.2 \
                 --squash

This adds two patches to the tree but these two patches do not build.

This patch fixes the build, I could not think of a way to keep the code
building on all commits.

Includes changes to the build script by Jake and update to the C++
toolchain suggested by Richard.

Uses the new version format, increasing the Bitocoin Core version as is
applicable and also bumping the minor version of this lib.

Co-developed-by: Jake Rawsthorne <jake@jakerawsthorne.co.uk>
Co-developed-by: Richard Ulrich <richard.ulrich@seba.swiss>