Change the crate version format
Currently we use a version string of the form `0.20.2-0.6.1`

While attempting to release of the `0.6` we found out that cargo ignores
the stuff after the `-` so the newly released code is treated as a
non-breaking change and pulled immediately by any crate that has
`bitcoinconsensus` in its dependency graph. This broke everyone that
uses it including `rust-bitcoin`.

We have since yanked both the `0.6` releases.

In order to function correctly with `cargo` change the format to put the
crate version first and put the Core version second, note we use
`+` (build) instead of `-` (pre-release). This should make `crates.io`
display the latest version correctly also.