CI: Add script and update github actions
Add a `test.sh` script similarly to how we do in `rust-secp256k1`, use
it to control running the other `_test.sh` script with the two manifest
files.

Modify the github action to use the test script, also update the runner
and set up jobs as we do in `rust-secp256k1`.